requirements.txt
discord.py == 1.7.3
PyNaCl == 1.4.0
Pillow == 8.3.1
numpy == 1.21.2
bs4 == 0.0.1
markdown == 3.3.4
discord-components == 2.0.4